:fire: delete useless banner
