+++ feature/main/src/main/java/com/msg/main/component/ClubBanner.kt
@@ -1,4 +1,4 @@
-package com.msg.main.banner.club
+package com.msg.main.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -12,14 +12,10 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.msg.design_system.theme.BitgoeulAndroidTheme
-import com.msg.main.component.AutoSchoolClubGridView
-import com.msg.main.component.getSpecificFieldList
 import com.msg.model.entity.school.GetSchoolListEntity
 import com.msg.model.enumdata.Field
-import com.msg.ui.DevicePreviews
 
 @Composable
 internal fun ClubBanner(
